clinigraph-ai/api/requirements-agentai.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
langchain
langgraph
langchain-openai
langchain-anthropic
langchain-ollama
langchain-huggingface
openai
anthropic
pinecone
weaviate-client
redis
kafka-python
python-dotenv
psycopg2-binary
stripe
reportlab
